// src/pages/FirstTeam/types.ts
export interface InputField {
    teamName: string;
    teamStatus: boolean;
    departmentRefId: number;
    teamRefId: number;
    error: string;
    organisationId: number;
    departmentId: number;
}

export interface StyledLottieProps {
    width?: string;
    paddingTop?: string;
    marginBottom?: string;
}